requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
jinja2==3.1.4
openai-whisper==20250625
gTTS==2.5.3
pydub==0.25.1
python-dotenv==1.0.1
aiofiles==24.1.0
rapidfuzz==3.10.1
